Update coercion failure handling (#22)
Change the behavior when a coercion failure occurs. Previously a failure will result in a nil value and will result in a hard error (It did not respect sorbet configuration).

So instead, we should return the original value when coercion fails. The environment is configured to trigger a hard error, it works as before (raised TypeError from `T.let`). If the environment is configured for soft errors only, now we will use the original value instead of failing it by force.