Avoid runtime method definitions (#42)
### Motivation
In ruby-prof reports, redefining methods in runtime creates quite a bit of overhead, especially in the situations where we need to coerce in batches. It'd be nice if the code can run faster.

### Approach
This adds a proxy-like class to handle the `.new` call. What happens under the hood is it's calling new on an instance, instead of `Class.new`. We hide this implementation from sorbet so it still works as before. This also removes the sigs on the method definitions.

### Test Plan
The tests cover the introduced changes. Generate the ruby-prof report again with the new code, see a ~30% performance improvement.

Pull Request Link: https://github.com/chanzuckerberg/sorbet-coerce/pull/42
Pull Request Branch: donaldong/avoid_runtime_method_definitions